refactor: remove common mark dir
