+++ src/mark/annotation/index.ts
@@ -1,4 +1,1 @@
-export {
-  Text as AnnotationText,
-  TextOptions as AnnotationTextOptions,
-} from './text';
+export { AnnotationText, AnnotationTextOptions } from './text';
