+++ src/mark/geometry/line.ts
@@ -3,7 +3,7 @@ import { group } from 'd3-array';
 import { isParallel } from '../../utils/coordinate';
 import { Mark, MarkComponent as MC, Vector2 } from '../../runtime';
 import { LineGeometry } from '../../spec';
-import { baseChannels, baseInference } from '../common/utils';
+import { baseChannels, baseInference } from '../utils';
 
 export type LineOptions = Omit<LineGeometry, 'type'>;
