+++ src/mark/geometry/text.ts
@@ -1,10 +1,21 @@
-import { MarkComponent as MC } from '../../runtime';
-import { baseChannels, baseInference } from '../common/utils';
-import { PointMark } from '../common/text';
+import { MarkComponent as MC, Vector2 } from '../../runtime';
+import { baseChannels, baseInference } from '../utils';
 import { TextGeometry } from '../../spec';
 
 export type TextOptions = Omit<TextGeometry, 'type'>;
-export const Text: MC<TextOptions> = () => PointMark;
+export const Text: MC<TextOptions> = () => {
+  return (index, scale, value, coordinate) => {
+    const { x: X, y: Y } = value;
+    const { x } = scale;
+    const P = Array.from(index, (i) => {
+      const px = +X[i][0];
+      const py = +Y[i][0];
+      const xoffset = x?.getBandWidth?.() || 0;
+      return [coordinate.map([px + xoffset / 2, py])] as Vector2[];
+    });
+    return [index, P];
+  };
+};
 
 Text.props = {
   defaultShape: 'text',
